backend/requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
jwt==1.3.1
packaging==23.2
pycparser==2.21
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
